fix: rename gRPC docs files to use underscores matching Makefile targets
The generator and test files used hyphens (generate-grpc-docs.py) but the
Makefile referenced them with underscores (generate_grpc_docs.py), causing
docs-check-grpc, docs-test-grpc, and CI documentation workflow to fail.

Renames to underscore convention (Python module standard), updates
multiversion.sh, and replaces fragile importlib.util import with standard
import statement.

Generated-By: Forge/20260526_163342_3065934_106c7b0c
